fix: style clippy

+++ src/utils/geom.rs
@@ -6,6 +6,6 @@ pub fn calc_dist_two_points(p1: &(i32, i32), p2: &(i32, i32)) -> f32 {
 
 #[test]
 fn test_calc_dist_two_points() {
-    assert_eq!(calc_dist_two_points(&(1, 1), &(2, 2)), (2 as f32).powf(0.5));
-    assert_eq!(calc_dist_two_points(&(1, 0), &(4, 4)), 5 as f32);
+    assert_eq!(calc_dist_two_points(&(1, 1), &(2, 2)), (2_f32).powf(0.5));
+    assert_eq!(calc_dist_two_points(&(1, 0), &(4, 4)), 5_f32);
 }
